Obliterate & flatten the settings class tree.
This gets rid of a lot of redundant classes in-between Setting & AbstractSetting.
