+++ src/main/java/me/zeroeightsix/kami/command/commands/SetCommand.java
@@ -6,7 +6,6 @@ import me.zeroeightsix.kami.command.syntax.parsers.ModuleParser;
 import me.zeroeightsix.kami.module.Module;
 import me.zeroeightsix.kami.module.ModuleManager;
 import me.zeroeightsix.kami.setting.ISettingUnknown;
-import me.zeroeightsix.kami.setting.Named;
 import me.zeroeightsix.kami.setting.Setting;
 
 import java.util.Optional;
@@ -39,7 +38,7 @@ public class SetCommand extends Command {
         }
 
         if (args[1] == null) {
-            String settings = String.join(", ",  m.settingList.stream().filter(setting -> setting instanceof Named).map(setting -> ((Named) setting).getName()).collect(Collectors.toList()));
+            String settings = String.join(", ",  m.settingList.stream().map(setting -> setting.getName()).collect(Collectors.toList()));
             if (settings.isEmpty())
                 Command.sendChatMessage("Module &b" + m.getName() + "&r has no settings.");
             else {
@@ -50,7 +49,7 @@ public class SetCommand extends Command {
             return;
         }
 
-        Optional<Setting> optionalSetting = m.settingList.stream().filter(setting1 -> setting1 instanceof Named).filter(setting1 -> ((Named) setting1).getName().equalsIgnoreCase(args[1])).findFirst();
+        Optional<Setting> optionalSetting = m.settingList.stream().filter(setting1 -> setting1.getName().equalsIgnoreCase(args[1])).findFirst();
         if (!optionalSetting .isPresent()) {
             Command.sendChatMessage("Unknown setting &b" + args[1] + "&r in &b" + m.getName() + "&r!");
             return;
@@ -59,13 +58,13 @@ public class SetCommand extends Command {
         ISettingUnknown setting = optionalSetting.get();
 
         if (args[2] == null) {
-            Command.sendChatMessage("&b" + ((Named) setting).getName() + "&r is a &3" + setting.getValueClass().getSimpleName() + "&r. Its current value is &3" + setting.getValueAsString());
+            Command.sendChatMessage("&b" + setting.getName() + "&r is a &3" + setting.getValueClass().getSimpleName() + "&r. Its current value is &3" + setting.getValueAsString());
             return;
         }
 
         try {
             setting.setValueFromString(args[2]);
-            Command.sendChatMessage("Set &b" + ((Named) setting).getName() + "&r to &3" + args[2] + "&r.");
+            Command.sendChatMessage("Set &b" + setting.getName() + "&r to &3" + args[2] + "&r.");
         } catch (Exception e) {
             e.printStackTrace();
             Command.sendChatMessage("Unable to set value! &6" + e.getMessage());
