--- src/main/java/me/zeroeightsix/kami/setting/Named.java
@@ -1,10 +1,0 @@
-package me.zeroeightsix.kami.setting;
-
-/**
- * Created by 086 on 12/10/2018.
- */
-public interface Named {
-
-    String getName();
-
-}
